add dark mode and supports ipados

+++ toDoList/view/ListView.swift
@@ -15,36 +15,52 @@ struct ListView: View {
     
     var body: some View {
         
-        NavigationView(){
-            List{
-                ForEach(listViewModel.items) { item in
-                    listRowView(item: item)
-                        .onTapGesture {
-                            withAnimation(.linear){
-                                listViewModel.updateItem(item: item)
-                                
-                            }
-                        }
-                }
-                .onDelete(perform: listViewModel.deleteItem)
-                .onMove(perform: listViewModel.moveItem)
+        ZStack{
+            //если список пустой
+            if listViewModel.items.isEmpty {
                 
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.linear))
             }
-            .navigationBarItems(
-                leading: EditButton(),
-                trailing:
-                    NavigationLink("Add",destination: addNotsView())
-            )
-            .navigationTitle("Todo List")
+            else {
+                List{
+                    ForEach(listViewModel.items) { item in
+                        listRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item: item)
+                                    
+                                }
+                            }
+                    }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                    
+                }
+            }
+            
         }
+        
+        
+        .navigationBarItems(
+            leading: EditButton(),
+            trailing:
+                NavigationLink("Add",destination: addNotsView())
+        )
+        .navigationTitle("Todo List")
+        
     }
+    
     
 }
 
 struct ListView_Previews: PreviewProvider {
     static var previews: some View {
         
-        ListView()
-            .environmentObject(ListViewModels())
+        NavigationView {
+            ListView()
+        }
+        .environmentObject(ListViewModels())
+        .navigationViewStyle(StackNavigationViewStyle())
     }
 }
